Time: 0 ms (100.00%) | Memory: 8.7 MB (3.89%) - LeetSync

+++ 2278-find-three-consecutive-integers-that-sum-to-a-given-number/find-three-consecutive-integers-that-sum-to-a-given-number.cpp
@@ -4,7 +4,7 @@ public:
     vector<long long> ans;
 
     if(num % 3 == 0){
-        long long x = num/3;
+        long x = num/3;
         ans.push_back(x-1);
         ans.push_back(x);
         ans.push_back(x+1);
